+++ CHANGES.md
@@ -3,9 +3,13 @@
 ## 0.19.1 (2026-xx-xx)
 
 * Fix [Issue #134](https://github.com/julien6387/supvisors/issues/134).
-  Reload the Supvisors process configuration after Supervisor is updated.
+  Reload the **Supvisors** process configuration after Supervisor is updated.
 
-* Strengthen robustness against `psutil` issues realted to certain operating systems.
+* Set the lower limit of the option `synchro_timeout` to 0 instead of 15 (Default value remains unchanged).
+
+* Speed-up the entry in SYNCHRONIZATION state.
+
+* Strengthen robustness against `psutil` issues related to certain operating systems.
 
 * Review HTTP error codes returned by the REST API in `supvisorsflask`.